requirements.txt
ipykernel
annoy==1.17.3
numpy==1.26.4
ultralytics==8.3.12
fastapi[all]==0.115.11
faiss-cpu==1.10.0
from_root
Pillow
pyyaml
uuid
streamlit>=1.32.0
matplotlib>=3.8.0
opencv-python>=4.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
plotly>=5.18.0
-e .